Add supported parameters to GetAccountTariffsRequest.java
Currently, parameters defined in this class, such as populateRates, are not propagated to the Account Service. As a result, the HTTP request sent by this class does not include useful or supported API query parameters.

This change ensures that these parameters are correctly added to the request, allowing the Account Service to receive and process them as intended.